affichage du perso NEWS

+++ src/draw/cub_render.c
@@ -6,7 +6,7 @@
 /*   By: ybouroga <ybouroga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/28 17:53:25 by ybouroga          #+#    #+#             */
-/*   Updated: 2025/11/28 12:33:45 by ybouroga         ###   ########.fr       */
+/*   Updated: 2025/11/28 14:34:15 by ybouroga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ static int	cub_draw_v_wall(t_cub *m, int x)
 	while (y <= m->rec.draw_end)
 	{
 		texture_position += step;
-		color = get_texture_pixel(texture, m->rec.texture_x, (int)texture_position);
+		color = get_texture_pixel(texture, m->rec.texture_x, /*(int)*/texture_position);
 		cub_putpixel(m, x, y, color);
 		y++;
 	}
